# Reduce memory usage of inserts to JSON by using adaptive write buffer size