--- a/src/Compression/CompressedWriteBuffer.h
+++ b/src/Compression/CompressedWriteBuffer.h
@@ -54,7 +54,7 @@ private:
 
     /// If true, the size of internal buffer will be exponentially increased up to
     /// adaptive_buffer_max_size after each nextImpl call. It can be used to avoid
-    /// large buffer allocation when actual size of writen data is small.
+    /// large buffer allocation when actual size of written data is small.
     bool use_adaptive_buffer_size;
     size_t adaptive_buffer_max_size;