--- a/src/Disks/IO/WriteBufferFromAzureBlobStorage.cpp
+++ b/src/Disks/IO/WriteBufferFromAzureBlobStorage.cpp
@@ -251,10 +251,27 @@ void WriteBufferFromAzureBlobStorage::allocateBuffer()
     auto size = buffer_allocation_policy->getBufferSize();
 
     if (buffer_allocation_policy->getBufferNumber() == 1)
-        size = std::min(size_t(DBMS_DEFAULT_BUFFER_SIZE), size);
+    {
+        allocateFirstBuffer();
+        return;
+    }
 
     memory = Memory(size);
     WriteBuffer::set(memory.data(), memory.size());
+}
+
+void WriteBufferFromAzureBlobStorage::allocateFirstBuffer()
+{
+    /// First buffer was already allocated in BufferWithOwnMemory constructor with buffer size provided in constructor.
+    /// It will be reallocated in subsequent nextImpl calls up to the desired buffer size from buffer_allocation_policy.
+    /// But it may happen that buffer size provided in constructor is larger then desired buffer size from buffer_allocation_policy.
+    /// Resize memory in this case to the desired size.
+    const auto max_first_buffer = buffer_allocation_policy->getBufferSize();
+    if (memory.size() > max_first_buffer)
+    {
+        memory.resize(max_first_buffer);
+        WriteBuffer::set(memory.data(), memory.size());
+    }
 }
 
 void WriteBufferFromAzureBlobStorage::detachBuffer()