--- a/src/Disks/IO/WriteBufferFromAzureBlobStorage.h
+++ b/src/Disks/IO/WriteBufferFromAzureBlobStorage.h
@@ -53,6 +53,7 @@ private:
     void detachBuffer();
     void reallocateFirstBuffer();
     void allocateBuffer();
+    void allocateFirstBuffer();
     void hidePartialData();
     void setFakeBufferWhenPreFinalized();