--- a/src/Disks/IO/WriteBufferFromAzureBlobStorage.h
+++ b/src/Disks/IO/WriteBufferFromAzureBlobStorage.h
@@ -53,7 +53,6 @@ private:
     void detachBuffer();
     void reallocateFirstBuffer();
     void allocateBuffer();
-    void allocateFirstBuffer();
     void hidePartialData();
     void setFakeBufferWhenPreFinalized();